follow the ethereum-lcp-poc manner

+++ module/prover_test.go
@@ -129,7 +129,7 @@ func (ts *ProverTestSuite) SetupTest() {
 	config := ProverConfig{
 		TrustLevelNumerator:   1,
 		TrustLevelDenominator: 3,
-		TrustingPeriod:        time.Second * 100,
+		TrustingPeriod:        uint64(time.Second * 100),
 	}
 	ts.chain = &mockChain{
 		Chain:        NewChain(chain),
@@ -202,7 +202,7 @@ func (ts *ProverTestSuite) TestCreateMsgCreateClient() {
 	var cs ClientState
 	ts.Require().NoError(proto.Unmarshal(msg.ClientState.Value, &cs))
 	ts.Require().Equal(cs.ChainId, uint64(9999))
-	ts.Require().Equal(cs.TrustingPeriod, time.Second*100)
+	ts.Require().Equal(cs.TrustingPeriod, uint64(time.Second*100))
 	ts.Require().Equal(cs.TrustLevel.Numerator, uint64(1))
 	ts.Require().Equal(cs.TrustLevel.Denominator, uint64(3))
 	ts.Require().False(cs.Frozen)
